refactored where example optimization problems go
